Allow to directly connect callmonitor and contact node
This enhances the contact (lookup) node so that the callmonitor can
directly be connected as input. In this case it adds caller_contacts
and callee_contacts arrays with the contact information from the
phonebook.

In addition this changes the error behavior so that the node returns an
empty array in case of phonenumber parsing errors.